% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{.meta_unique}
\alias{.meta_unique}
\title{Return a single value from a metadata column, collapsing multiple values}
\usage{
.meta_unique(meta, col)
}
\description{
Return a single value from a metadata column, collapsing multiple values
}
\keyword{internal}
