% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{sc_annotate_manual}
\alias{sc_annotate_manual}
\title{Apply manual cluster-to-celltype annotation}
\usage{
sc_annotate_manual(merged, annotation, cfg, new_col = "celltype_manual")
}
\arguments{
\item{merged}{Seurat object.}

\item{annotation}{Named character vector: \code{c("1" = "T cells", ...)}.}

\item{cfg}{Config list from \code{\link[=sc_config]{sc_config()}}.}

\item{new_col}{Name of the new metadata column.}
}
\value{
Seurat object with manual annotation column added.
}
\description{
Apply manual cluster-to-celltype annotation
}
