% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{sc_cache_clear}
\alias{sc_cache_clear}
\title{Delete one or all cache files}
\usage{
sc_cache_clear(cache_dir, cache_file = NULL)
}
\arguments{
\item{cache_dir}{Cache directory path.}

\item{cache_file}{Optional. Specific file name to delete. If \code{NULL},
all cache files are deleted.}
}
\description{
Delete one or all cache files
}
