% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{sc_cache_dir}
\alias{sc_cache_dir}
\title{Return the cache directory from a config object}
\usage{
sc_cache_dir(cfg)
}
\arguments{
\item{cfg}{Config list returned by \code{\link[=sc_config]{sc_config()}}.}
}
\value{
Character path.
}
\description{
Return the cache directory from a config object
}
