% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{sc_cache_list}
\alias{sc_cache_list}
\title{List all cache files for a project}
\usage{
sc_cache_list(cache_dir)
}
\arguments{
\item{cache_dir}{Cache directory path.}
}
\value{
Character vector of \code{.qs2} file paths.
}
\description{
List all cache files for a project
}
