% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{sc_cluster}
\alias{sc_cluster}
\title{Cluster cells at multiple resolutions}
\usage{
sc_cluster(merged, cfg, force = FALSE)
}
\arguments{
\item{merged}{Integrated Seurat object from \code{\link[=sc_integrate]{sc_integrate()}}.}

\item{cfg}{Config list from \code{\link[=sc_config]{sc_config()}}.}

\item{force}{If \code{TRUE}, ignore cache.}
}
\value{
Clustered Seurat object.
}
\description{
Runs FindNeighbors then FindClusters at all resolutions defined in
\code{params.yml}. Sets Idents to the default resolution.
}
