% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{sc_doublets}
\alias{sc_doublets}
\title{Run scDblFinder doublet detection on all samples}
\usage{
sc_doublets(obj_list, cfg, force = FALSE)
}
\arguments{
\item{obj_list}{Named list of Seurat objects.}

\item{cfg}{Config list from \code{\link[=sc_config]{sc_config()}}.}

\item{force}{If \code{TRUE}, ignore cache.}
}
\value{
Named list with \code{scDblFinder_score} and \code{scDblFinder_class} added.
}
\description{
Run scDblFinder doublet detection on all samples
}
