% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_plots.R
\name{sc_plot_annotation}
\alias{sc_plot_annotation}
\title{UMAP coloured by manual annotation}
\usage{
sc_plot_annotation(merged, annot_col = "celltype_manual")
}
\arguments{
\item{merged}{Seurat object with manual annotation.}

\item{annot_col}{Metadata column name (default: \code{"celltype_manual"}).}
}
\value{
A ggplot object.
}
\description{
UMAP coloured by manual annotation
}
