% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_plots.R
\name{sc_plot_azimuth_umap}
\alias{sc_plot_azimuth_umap}
\title{UMAP coloured by Azimuth predictions}
\usage{
sc_plot_azimuth_umap(merged, level = "l1")
}
\arguments{
\item{merged}{Annotated Seurat object.}

\item{level}{Azimuth annotation level: \code{"l1"} or \code{"l2"}.}
}
\value{
A patchwork ggplot.
}
\description{
UMAP coloured by Azimuth predictions
}
