% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{sc_plot_complexity}
\alias{sc_plot_complexity}
\title{Cell complexity density plot}
\usage{
sc_plot_complexity(obj_list, cfg, fill_by = NULL, facet_by = "sample_id")
}
\arguments{
\item{obj_list}{Named list of Seurat objects.}

\item{cfg}{Config list from \code{\link[=sc_config]{sc_config()}}.}

\item{fill_by}{Metadata column used for fill. Defaults to best available.}

\item{facet_by}{Metadata column used for faceting. Defaults to \code{"sample_id"}.}
}
\value{
A ggplot object.
}
\description{
Cell complexity density plot
}
