% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_plots.R
\name{sc_plot_doublets}
\alias{sc_plot_doublets}
\title{Bar chart: \% doublets per sample}
\usage{
sc_plot_doublets(obj_list, cfg, fill_by = NULL)
}
\arguments{
\item{obj_list}{Named list of Seurat objects after \code{\link[=sc_doublets]{sc_doublets()}}.}

\item{cfg}{Config list from \code{\link[=sc_config]{sc_config()}}.}

\item{fill_by}{Metadata column for fill colour. Defaults to best available.}
}
\value{
A ggplot object.
}
\description{
Bar chart: \% doublets per sample
}
