% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plots.R
\name{sc_plot_elbow}
\alias{sc_plot_elbow}
\title{Elbow plot for PC selection}
\usage{
sc_plot_elbow(merged, cfg, n_show = 50L)
}
\arguments{
\item{merged}{Seurat object after PCA.}

\item{cfg}{Config list from \code{\link[=sc_config]{sc_config()}}.}

\item{n_show}{Number of PCs to display.}
}
\value{
A ggplot object.
}
\description{
Elbow plot for PC selection
}
