% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plots.R
\name{sc_plot_hvg}
\alias{sc_plot_hvg}
\title{Highly variable genes plot}
\usage{
sc_plot_hvg(merged, n_label = 20L)
}
\arguments{
\item{merged}{Normalised Seurat object.}

\item{n_label}{Number of top HVGs to label.}
}
\value{
A ggplot object.
}
\description{
Handles both SCTransform v2 and LogNormalize workflows automatically.
}
