% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{sc_plot_qc_distri}
\alias{sc_plot_qc_distri}
\title{Strip plot of median UMI and gene counts per sample}
\usage{
sc_plot_qc_distri(obj_list, cfg)
}
\arguments{
\item{obj_list}{Named list of Seurat objects.}

\item{cfg}{Config list from \code{\link[=sc_config]{sc_config()}}.}
}
\value{
A ggplot object.
}
\description{
Strip plot of median UMI and gene counts per sample
}
