% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_plots.R
\name{sc_plot_singler_heatmap}
\alias{sc_plot_singler_heatmap}
\title{SingleR score heatmap per cluster}
\usage{
sc_plot_singler_heatmap(merged, ref_name, cfg)
}
\arguments{
\item{merged}{Annotated Seurat object.}

\item{ref_name}{Reference name.}

\item{cfg}{Config list from \code{\link[=sc_config]{sc_config()}}.}
}
\value{
A pheatmap object, or \code{NULL}.
}
\description{
SingleR score heatmap per cluster
}
