% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_plots.R
\name{sc_plot_singler_umap}
\alias{sc_plot_singler_umap}
\title{UMAP coloured by SingleR labels}
\usage{
sc_plot_singler_umap(merged, ref_name)
}
\arguments{
\item{merged}{Annotated Seurat object.}

\item{ref_name}{Reference name used in SingleR (e.g. "HumanPrimaryCellAtlasData").}
}
\value{
A patchwork ggplot, or \code{NULL} if columns not found.
}
\description{
UMAP coloured by SingleR labels
}
