% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{sc_prefilter}
\alias{sc_prefilter}
\title{Apply a permissive pre-filter to remove true empty droplets}
\usage{
sc_prefilter(obj)
}
\arguments{
\item{obj}{A Seurat object.}
}
\value{
Filtered Seurat object.
}
\description{
Retains cells with at least 100 genes and 200 UMIs. This low threshold
avoids biasing ambient RNA estimation by decontX.
}
