% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{sc_qc_all_thresholds}
\alias{sc_qc_all_thresholds}
\title{Get thresholds for all samples}
\usage{
sc_qc_all_thresholds(cfg)
}
\arguments{
\item{cfg}{Config list from \code{\link[=sc_config]{sc_config()}}.}
}
\value{
Named list of threshold lists, one per sample.
}
\description{
Get thresholds for all samples
}
