% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{sc_qc_thresholds_df}
\alias{sc_qc_thresholds_df}
\title{Build a data frame of thresholds (for plotting)}
\usage{
sc_qc_thresholds_df(thresholds_list, cfg)
}
\arguments{
\item{thresholds_list}{Output of \code{\link[=sc_qc_all_thresholds]{sc_qc_all_thresholds()}}.}

\item{cfg}{Config list.}
}
\value{
A tibble with one row per sample, including override flag.
}
\description{
Build a data frame of thresholds (for plotting)
}
