% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{sc_sample_meta}
\alias{sc_sample_meta}
\title{Return sample metadata as a data frame}
\usage{
sc_sample_meta(cfg)
}
\arguments{
\item{cfg}{Config list returned by \code{\link[=sc_config]{sc_config()}}.}
}
\value{
A data frame with one row per sample.
}
\description{
Return sample metadata as a data frame
}
