% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plots.R
\name{sc_tabset}
\alias{sc_tabset}
\title{Render a named list of plots as a Quarto panel-tabset}
\usage{
sc_tabset(plot_list, print_fn = print)
}
\arguments{
\item{plot_list}{Named list of ggplot objects.}

\item{print_fn}{Function used to render each plot (default: \code{print}).}
}
\description{
Must be called inside a chunk with \code{results='asis'}.
}
