% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scReportR-package.R
\docType{package}
\name{scReportR-package}
\alias{scReportR}
\alias{scReportR-package}
\title{scReportR: Single-Cell RNAseq Analysis Report Generator}
\description{
Provides a standardised workflow for single-cell RNAseq analysis.
Use \code{\link[=init_sc_report]{init_sc_report()}} to initialise a new project, edit \code{params.yml},
then render \code{analysis.qmd} section by section.
}
\author{
\strong{Maintainer}: Your Name \email{your@email.com}

}
\keyword{internal}
