% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plots.R
\name{sc_plot_markers_dot}
\alias{sc_plot_markers_dot}
\title{DotPlot of known markers across clusters}
\usage{
sc_plot_markers_dot(merged, markers_list, cfg)
}
\arguments{
\item{merged}{Seurat object.}

\item{markers_list}{Named list of character vectors.}

\item{cfg}{Config list from \code{\link[=sc_config]{sc_config()}}.}
}
\value{
A plot object.
}
\description{
DotPlot of known markers across clusters
}
